image-cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-agent"
version = "0.1.0"
description = "AI image generation CLI tool with FLUX.2 support"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "torch",
    "diffusers",
    "transformers",
    "Pillow",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
image = "image_entry:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["common*", "cli*", "core*", "plugins*", "commands*", "api*", "image_entry*"]
